tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/089.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: async attribute on inline script</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script async>
        var t = async_test();
        log('inline script #1');
	t.step(function() {
                   assert_array_equals(eventOrder, ['inline script #1']);
               });
        </script>
	<script async>
        log('inline script #2');
	t.step(function() {
                   assert_array_equals(eventOrder, ['inline script #1', 'inline script #2']);
               });
        </script>

	<script>
        log('inline script #3');
	t.step(function() {
                   assert_array_equals(eventOrder, ['inline script #1', 'inline script #2', 'inline script #3']);
               });
        onload = function() {t.done()};
	</script>

</body></html>
